pyproject.toml
[tool.poetry]
name = "neuralcodr"
version = "0.0.1"
authors = ["Shroominic <contact@shroominic.com>"]
description = "Neuralcodr is an ai tool for faster coding."

[tool.poetry.dependencies]
python = "^3.11"
langchain = "*"
openai = "*"
python-dotenv = "*"
typer = "^0.9.0"
tiktoken = "^0.4.0"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "^3.3.3"
mypy = "^0.991"
black = "^22.12.0"
isort = "^5.11.2"
autoflake = "^2.0.0"
flake8 = "^6.0.0"
pytest = "^7.2.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
